services/companies-sync/requirements.txt
psycopg2-binary
fastapi
uvicorn
pika